+++ users/models/users.py
@@ -1,20 +1,10 @@
-from django.db import models
 from django.contrib.auth.models import AbstractUser
+from django.db import models
+
+from utils.choices.choice import CurrencyChoices, GenderChoices, LanguageChoices
 
 
 class User(AbstractUser):
-    class GenderChoices(models.TextChoices):
-        MALE = ("male", "Male")
-        FEMALE = ("female", "Female")
-
-    class LanguageChoices(models.TextChoices):
-        KR = ("kr", "Korean")
-        EN = ("en", "English")
-
-    class CurrencyChoices(models.TextChoices):
-        WON = "won", "Korean Won"
-        USD = "usd", "Dollar"
-
     first_name = models.CharField(
         max_length=150,
         editable=False,
