+++ utils/choices/choice.py
@@ -15,3 +15,18 @@ class RoomKindChoices(models.TextChoices):
     ENTIRE_PLACE = ("entire_place", "Entire Place")
     PRIVATE_ROOM = ("private_room", "Private Room")
     SHARED_ROOM = "shared_room", "Shared Room"
+
+
+class GenderChoices(models.TextChoices):
+    MALE = ("male", "Male")
+    FEMALE = ("female", "Female")
+
+
+class LanguageChoices(models.TextChoices):
+    KR = ("kr", "Korean")
+    EN = ("en", "English")
+
+
+class CurrencyChoices(models.TextChoices):
+    WON = "won", "Korean Won"
+    USD = "usd", "Dollar"
